+++ src/main/scala/leo/modules/output/LPoutput/LPSignature.scala
@@ -48,8 +48,8 @@ object LPSignature {
   }
 
   case object lpEm extends lpAxioms{
-    override def name: lpConstantTerm = lpConstantTerm("em")
     //  Π x: Prop, Prf (x ∨ ¬ x)
+    override def name: lpConstantTerm = lpConstantTerm("em")
     override def ty: lpMlType = lpMlDependType(Seq(lpTypedVar(lpConstantTerm("x"),lpOtype.lift2Meta)),lpOlUntypedBinaryConnectiveTerm(lpOr,lpOlConstantTerm("x"),lpOlUnaryConnectiveTerm(lpNot,lpOlConstantTerm("x"))).prf)
     override def pretty: String = lpDeclaration(lpEm.name,Seq.empty,lpEm.ty).pretty
   }
@@ -57,7 +57,7 @@ object LPSignature {
   case object lpPropExt extends lpAxioms {
     override def name: lpConstantTerm = lpConstantTerm("propExt")
 
-    //                                                                                                                Π x: Els (↑ o), Π y: Els (↑ o), (Prf x → Prf y) → (Prf y → Prf x) → Prf (eq x y)
+    // Π x: Els (↑ o), Π y: Els (↑ o), (Prf x → Prf y) → (Prf y → Prf x) → Prf (eq x y)
     override def ty: lpMlType = lpMlDependType(Seq(lpTypedVar(lpConstantTerm("x"), lpOtype.lift2Meta),lpTypedVar(lpConstantTerm("y"), lpOtype.lift2Meta)), lpMlFunctionType(Seq(lpMlFunctionType(Seq(lpOlConstantTerm("x").prf,lpOlConstantTerm("y").prf)),lpMlFunctionType(Seq(lpOlConstantTerm("y").prf,lpOlConstantTerm("x").prf)),lpOlTypedBinaryConnectiveTerm(lpEq,lpOtype,lpOlConstantTerm("x"),lpOlConstantTerm("y")).prf)))
 
     override def pretty: String = lpDeclaration(lpPropExt.name, Seq.empty, lpPropExt.ty).pretty
