+++ src/main/scala/leo/modules/output/LPoutput/lpInferenceRuleEncoding.scala
@@ -1,9 +1,13 @@
 package leo.modules.output.LPoutput
 
-import leo.modules.output.LPoutput.LPSignature.{HOLtop, Prf, lpPropExt, oType,lpEm}
-import leo.modules.output.LPoutput.calculusEncoding.{nameBottom, nameHypothesis, nameX, nameType}
+import leo.modules.output.LPoutput.LPSignature._
 import leo.modules.output.LPoutput.lpDatastructures._
-import leo.modules.output.LPoutput.lpUseful._
+import leo.modules.output.LPoutput.NaturalDeductionRules._
+
+/** Definitions of the Inferences rules of the calculus EP
+  *
+  * @author Melanie Taprogge
+  */
 
 object lpInferenceRuleEncoding {
 
@@ -13,17 +17,18 @@ object lpInferenceRuleEncoding {
 
   }
 
+  ////////////////////////////////////////////////////////////////
+  ////////// Primary Inference Rules
+  ////////////////////////////////////////////////////////////////
 
   case class eqFactoring_script(polarity: Boolean) extends inferenceRules {
-
     // pos: [T] x y z v : Prf((eq [T] x y) ∨ (eq [T] z v)) → Prf((eq [T] x y) ∨ ¬ (eq [T] x z) ∨ ¬ (eq [T] y v))
     // neg: [T] x y z v : Prf(¬ (eq [T] x y) ∨ ¬ (eq [T] z v)) → Prf(¬ (eq [T] x y) ∨ ¬ (eq [T] x z) ∨ ¬ (eq [T] y v))
+
     override def name: lpConstantTerm = {
       val pol = if (polarity) "_p" else "_n"
       lpConstantTerm(s"EqFact$pol")
     }
-
-    //def res(T0, x0, y0, z0, v0):lpOlTerm = {}
 
     val T = lpOlUserDefinedPolyType("T")
     val x = lpOlConstantTerm("x")
@@ -56,27 +61,29 @@ object lpInferenceRuleEncoding {
         Seq(lpOlUnaryConnectiveTerm(lpNot, lpOlTypedBinaryConnectiveTerm(lpEq, T0, x0, y0)), lpOlUnaryConnectiveTerm(lpNot, lpOlTypedBinaryConnectiveTerm(lpEq, T0, x0, z0)), lpOlUnaryConnectiveTerm(lpNot, lpOlTypedBinaryConnectiveTerm(lpEq, T0, y0, v0)))
       }
     }
-
   }
 
 
-  case class funextPosEq_rev(n: Int = 1) extends inferenceRules {
+  ////////////////////////////////////////////////////////////////
+  ////////// Extensionality
+  ////////////////////////////////////////////////////////////////
+
+  case class funExtPosEq_rev() extends inferenceRules {
+    // [T S] f g x : Prf(= (= [S] (f x) (g x)) (= [T ⤳ S] f g))
 
     override val proofIsDefined = true
 
-    if (n > 1) throw new Exception(s"error encoding funcExt: trying to instanciate rule for more than one arg not encoded yet")
     override def name: lpConstantTerm = lpConstantTerm(s"PFE")
 
-    // [T S] f g x : Prf(= (= [S] (f x) (g x)) (= [T ⤳ S] f g))
     val T = lpOlUserDefinedMonoType("T")
     val S = lpOlUserDefinedMonoType("S")
     val f = lpOlConstantTerm("f")
     val g = lpOlConstantTerm("g")
-    val x = lpOlConstantTerm("x") //todo i need n x here
+    val x = lpOlConstantTerm("x")
+
     override def ty: lpMlType = lpOlTypedBinaryConnectiveTerm(lpEq, lpOtype, lpOlTypedBinaryConnectiveTerm(lpEq,S,lpOlFunctionApp(f,Seq(x)),lpOlFunctionApp(g,Seq(x))),lpOlTypedBinaryConnectiveTerm(lpEq,lpOlFunctionType(Seq(T,S)),f,g)).prf
 
-    // todo: encode properly and generate variably
-    override def proof: lpProofScript = lpProofScript(Seq(lpProofScriptCommentLine("this is statically encoded, change that...\n    assume T S f g x;\n\n    have H1: Prf(= [S] (f x) (g x)) → Prf(= [(T ⤳ S)] f g)\n        {refine (funExt [T] [S] f g) x};\n\n    have H2: Prf((= [(T ⤳ S)] f g)) → Prf(= [S] (f x) (g x))\n        {assume h;\n        refine =def [(T ⤳ S)] f g h (λ y, = [S] (y x) (g x)) (=ref [S] (g x))};\n\n    refine (propExt (= [S] (f x) (g x)) (= [(T ⤳ S)] f g)) H1 H2;\n"))) //todo: generate depending on number of args
+    override def proof: lpProofScript = lpProofScript(Seq(lpProofScriptCommentLine("    assume T S f g x;\n\n    have H1: Prf(= [S] (f x) (g x)) → Prf(= [(T ⤳ S)] f g)\n        {refine (funExt [T] [S] f g) x};\n\n    have H2: Prf((= [(T ⤳ S)] f g)) → Prf(= [S] (f x) (g x))\n        {assume h;\n        refine =def [(T ⤳ S)] f g h (λ y, = [S] (y x) (g x)) (=ref [S] (g x))};\n\n    refine (propExt (= [S] (f x) (g x)) (= [(T ⤳ S)] f g)) H1 H2;\n"))) //todo: generate depending on number of args
 
     override def usedBasicRules: Set[lpStatement] = Set(eqRef(),eqDef())
 
@@ -93,57 +100,9 @@ object lpInferenceRuleEncoding {
     }
   }
 
-  /*
-  case class funextPosEq_proofTerm(n: Int = 1, parameters0:(Int,Int,Int,Int)) extends inferenceRules {
-
-    if (n > 1) throw new Exception(s"error encoding eqFactoring: trying to instanciate rule for more than one arg not encoded yet")
-
-    var parameters = parameters0
-
-    override def name: lpConstantTerm = lpConstantTerm(s"funextPos${n}_proofTerm")
-
-    /*
-    // [T S] f g x : Prf(= (= [S] (f x) (g x)) (= [T ⤳ S] f g))
-    val T = lpOlUserDefinedPolyType("T")
-    val S = lpOlUserDefinedPolyType("S")
-    val f = lpOlConstantTerm("f")
-    val g = lpOlConstantTerm("g")
-    val x = lpOlConstantTerm("x") //todo i need n x here
-     */
-    val x1 = nameX(1)
-    val x2 = nameX(2)
-    val x3 = nameX(3)
-    val x4 = nameX(4)
-    val x5 = nameX(5)
-    val h1 = nameHypothesis(1)
-    val T1 = nameType(1)
-    val T2 = nameType(2)
-
-    // λ T S f g (h1 : Prf (eq [↑ (T ⤳ S)] f g)) a (p : Els (↑ S) → Prop), h1 (λ x, p (x a))
-    // Prf(eq [↑ (T ⤳ S)] f g) → Π a , Prf(eq [↑ S] (f a) (g a))
-    override def ty: lpMlType = lpMlDependType(Seq(lpUntypedVar(T1),lpUntypedVar(T2),lpUntypedVar(x3),lpUntypedVar(x4)))
-      //lpLambdaTerm(Seq(lpUntypedVar(T1),lpUntypedVar(T2),lpUntypedVar(x3),lpUntypedVar(x4),lpTypedVar(h1,lpOlTypedBinaryConnectiveTerm(lpEq,lpOlFunctionType(Seq(T1,T2)),x3,x4).prf),lpUntypedVar(x1),lpTypedVar(x2,T2)),lpOlFunctionApp(lpOlConstantTerm(h1.name),Seq(lpLambdaTerm(Seq(lpUntypedVar(x5)),lpOlFunctionApp(x2,Seq(lpFunctionApp(x5,Seq(x1))))))))
-
-    override def proof: lpProofScript = throw new Exception("proof for mkPosPropPosLit_script not encoded yet") //todo: generate depending on number of args
-
-    override def dec: lpDeclaration = lpDeclaration(name, Seq(f, g, x), ty, Seq(T, S))
-
-    override def pretty: String = lpDefinition(name, Seq(f, g, x), ty, proof, Seq(T, S)).pretty
-
-    def instanciate(TS0: Option[(lpOlPolyType, lpOlPolyType)], f: lpOlTerm, g: lpOlTerm, x: Seq[lpOlTerm]): lpFunctionApp = {
-      val ImpArgs = TS0 match {
-        case Some((t, s)) => Seq(t, s)
-        case None => Seq.empty
-      }
-      lpFunctionApp(name, Seq(f, g) ++ x, ImpArgs)
-    }
-  }
-*/
 
   case class boolExt(lhsNeg: Boolean, polarity: Boolean) extends inferenceRules {
-
-    // transform an equality literal of form a=b to either (¬ a ∨ b) or (a ∨ ¬ b)
-    // lhs encodes weather we want the negation on the lhs (¬ a ∨ b) or not, then we return the version for the rhs (a ∨ ¬ b)
+    // lhsNeg encodes weather we want the negation on the lhs
 
     override val proofIsDefined = true
 
@@ -179,16 +138,16 @@ object lpInferenceRuleEncoding {
 
     override def proof: lpProofScript = {
       if (polarity & !lhsNeg) {
-        lpProofScript(Seq(lpProofScriptCommentLine("this is not properly encoded yet, change that\n    assume x y h;\n    refine =def [o] x y h (λ z, z ∨ ¬ y) (em y);")))
+        lpProofScript(Seq(lpProofScriptCommentLine("    assume x y h;\n    refine =def [o] x y h (λ z, z ∨ ¬ y) (em y);")))
       } else if (polarity & lhsNeg) {
         // a b : Prf(eq [↑ o] a b) → Prf(a ∨ (¬ b))
-        lpProofScript(Seq(lpProofScriptCommentLine("this is not properly encoded yet, change that\n    assume x y h1;\n    have em_sym: Prf(¬ y ∨ y)\n        {refine ∨E y (¬ y) (¬ y ∨ y) _ _ (em y)\n            {assume h2;\n            refine ∨Ir (¬ y) y h2}\n            {assume h2;\n            refine ∨Il (¬ y) y h2}};\n    refine =def [o] x y h1 (λ z, ¬ z ∨ y) em_sym;")))
+        lpProofScript(Seq(lpProofScriptCommentLine("    assume x y h1;\n    have em_sym: Prf(¬ y ∨ y)\n        {refine ∨E y (¬ y) (¬ y ∨ y) _ _ (em y)\n            {assume h2;\n            refine ∨Ir (¬ y) y h2}\n            {assume h2;\n            refine ∨Il (¬ y) y h2}};\n    refine =def [o] x y h1 (λ z, ¬ z ∨ y) em_sym;")))
       } else if (!polarity & lhsNeg) {
         // x y: Prf(¬(= [o] x y)) → Prf(x ∨ y)
-        lpProofScript(Seq(lpProofScriptCommentLine("this is not properly encoded yet, change that\n    assume x y h1;\n   \n    refine ∨E x (¬ x) (x ∨ y) _ _ (em x)\n        {assume h2;\n        refine ∨Il x y h2}\n        {assume h2;\n        have yPrf: Prf y\n            {have notNotYprf: Prf (¬ y) → Prf ⊥\n                {assume h3;\n                have xImpY: Prf x → Prf y\n                    {assume h4;\n                    refine ⊥I y (¬E x h4 h2)};\n                have yImpX: Prf y → Prf x\n                    {assume h4;\n                    refine ⊥I x (¬E y h4 h3)};\n                refine (λ u, ¬E (= [o] x y) u h1) (propExt x y xImpY yImpX)};\n            refine npp (y) (¬I (¬ y) notNotYprf)};\n        refine ∨Ir x y yPrf};")))
+        lpProofScript(Seq(lpProofScriptCommentLine("    assume x y h1;\n   \n    refine ∨E x (¬ x) (x ∨ y) _ _ (em x)\n        {assume h2;\n        refine ∨Il x y h2}\n        {assume h2;\n        have yPrf: Prf y\n            {have notNotYprf: Prf (¬ y) → Prf ⊥\n                {assume h3;\n                have xImpY: Prf x → Prf y\n                    {assume h4;\n                    refine ⊥I y (¬E x h4 h2)};\n                have yImpX: Prf y → Prf x\n                    {assume h4;\n                    refine ⊥I x (¬E y h4 h3)};\n                refine (λ u, ¬E (= [o] x y) u h1) (propExt x y xImpY yImpX)};\n            refine npp (y) (¬I (¬ y) notNotYprf)};\n        refine ∨Ir x y yPrf};")))
       } else {
         // x y: Prf(¬(= [o] x y)) → Prf(¬ x ∨ ¬ y)
-        lpProofScript(Seq(lpProofScriptCommentLine("this is not properly encoded yet, change that\n    assume x y h1;\n   \n    refine ∨E x (¬ x) (¬ x ∨ ¬ y) _ _ (em x)\n        {assume h2;\n        have notNotYprf: Prf (y) → Prf ⊥\n            {assume h3;\n            have xImpY: Prf(x) → Prf(y)\n                {assume h4;\n                refine h3};\n            have yImpX: Prf(y) → Prf(x)\n                {assume h4;\n                refine h2};\n            refine (λ u, ¬E (= [o] x y) u h1) (propExt x y xImpY yImpX)};\n        refine ∨Ir (¬ x) (¬ y) (¬I y notNotYprf)}\n        {assume h2;\n        refine ∨Il (¬ x) (¬ y) h2}")))
+        lpProofScript(Seq(lpProofScriptCommentLine("    assume x y h1;\n   \n    refine ∨E x (¬ x) (¬ x ∨ ¬ y) _ _ (em x)\n        {assume h2;\n        have notNotYprf: Prf (y) → Prf ⊥\n            {assume h3;\n            have xImpY: Prf(x) → Prf(y)\n                {assume h4;\n                refine h3};\n            have yImpX: Prf(y) → Prf(x)\n                {assume h4;\n                refine h2};\n            refine (λ u, ¬E (= [o] x y) u h1) (propExt x y xImpY yImpX)};\n        refine ∨Ir (¬ x) (¬ y) (¬I y notNotYprf)}\n        {assume h2;\n        refine ∨Il (¬ x) (¬ y) h2}")))
       }
     }
 
@@ -203,6 +162,7 @@ object lpInferenceRuleEncoding {
         Set(lpEm, orE(),orIr(),orIl())
       }
     }
+
     override def dec: lpDeclaration = lpDeclaration(name, Seq(x,y), ty)
 
     override def pretty: String = lpDefinition(name, Seq(x,y), ty, proof).pretty
@@ -212,11 +172,16 @@ object lpInferenceRuleEncoding {
     }
   }
 
+
+  ////////////////////////////////////////////////////////////////
+  ////////// Additional Leo-III Inferences
+  ////////////////////////////////////////////////////////////////
+
   case object polaritySwitchEqLit extends inferenceRules {
+    // a b : Prf(= [o] (= [o] a b) (= [o] (¬ a) (¬ b)))
 
     override def name: lpConstantTerm = lpConstantTerm(s"polaritySwitchEqLit")
 
-    // a b : Prf(= [o] (= [o] a b) (= [o] (¬ a) (¬ b)))
     val a = lpOlConstantTerm("a")
     val b = lpOlConstantTerm("b")
 
@@ -233,26 +198,5 @@ object lpInferenceRuleEncoding {
     }
   }
 
-  /*
-  case object polaritySwitchNonEqLit extends inferenceRules {
-
-    override def name: lpConstantTerm = lpConstantTerm(s"polaritySwitchNonEqLit")
-
-    // a : Prf(= a (¬ ¬ a))
-    val a = lpOlConstantTerm("a")
-
-    override def ty: lpMlType = lpOlTypedBinaryConnectiveTerm(lpEq, lpOtype, a, lpOlUnaryConnectiveTerm(lpNot,lpOlUnaryConnectiveTerm(lpNot,a))).prf
-
-    override def proof: lpProofScript = throw new Exception("proof for mkPosPropPosLit_script not encoded yet") //todo: generate depending on number of args
-
-    override def dec: lpDeclaration = lpDeclaration(name, Seq(a), ty)
-
-    override def pretty: String = lpDefinition(name, Seq(a), ty, proof).pretty
-
-    def instanciate(a: lpOlTerm): lpFunctionApp = {
-      lpFunctionApp(name, Seq(a))
-    }
-  }
-   */
 
 }
